+++ src/main/org/firebirdsql/jca/FBManagedConnection.java
@@ -34,6 +34,7 @@ import javax.security.auth.Subject;
 import javax.transaction.xa.*;
 
 import org.firebirdsql.gds.*;
+import org.firebirdsql.gds.impl.DatabaseParameterBufferExtension;
 import org.firebirdsql.gds.impl.DbAttachInfo;
 import org.firebirdsql.gds.impl.GDSHelper;
 import org.firebirdsql.gds.impl.jni.EmbeddedGDSFactoryPlugin;
@@ -103,7 +104,8 @@ public class FBManagedConnection implements ManagedConnection, XAResource, Excep
         try {
             DatabaseParameterBuffer dpb = this.cri.getDpb();
 
-            if (dpb.getArgumentAsString(DatabaseParameterBuffer.LC_CTYPE) == null) {
+            if (dpb.getArgumentAsString(DatabaseParameterBuffer.LC_CTYPE) == null
+                    && dpb.getArgumentAsString(DatabaseParameterBufferExtension.LOCAL_ENCODING) == null) {
                 String defaultEncoding = getDefaultConnectionEncoding();
                 if (defaultEncoding == null) {
                     throw new SQLNonTransientConnectionException(ERROR_NO_CHARSET,